// src/app/@core/models/contrato_general.ts
export class ContratoGeneral {
    Id: string;
    VigenciaContrato: number;
    ObjetoContrato: string;
    PlazoEjecucion: number;
    FormaPago: any;
    OrdenadorGasto: number;
    ClausulaRegistroPresupuestal: boolean;
    SedeSolicitante: string;
    DependenciaSolicitante: string;
    Contratista: number;
    ValorContrato: number;
    Justificacion: string;
    DescripcionFormaPago: string;
    Condiciones: string;
    FechaRegistro: Date;
    TipologiaContrato: number;
    TipoCompromiso: number;
    ModalidadSeleccion: number;
    Procedimiento: number;
    RegimenContratacion: number;
    TipoGasto: number;
    TemaGastoInversion: number;
    OrigenPresupueso: number;
    OrigenRecursos: number;
    TipoMoneda: number;
    ValorContratoMe: number;
    ValorTasaCambio: number;
    TipoControl: number;
    Observaciones: string;
    ClaseContratista: number;
    Convenio: string;
    NumeroConstancia: number;
    Estado: boolean;
    UnidadEjecutora: number;
    UnidadEjecucion: any;
    LugarEjecucion: any;
    TipoContrato: any;
    OrigenPresupuesto: number;
}